requirements.txt
# Photo Organizer Dependencies
# Install with: pip install -r requirements.txt

Pillow>=9.0.0